Trim whitespace off command before storing
When using completion, commands often end up with a trailing space. Fish
itself appears to trim those spaces in its own history (couldn't confirm
it in the code though).

The spaces are not too much of an issue in themselves, but aligning this
behavior with fish mostly helps with diffing fishql and
builtin history for debugging purposes.